backend/requirements.txt
# FastAPI & Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database & ORM
asyncpg>=0.29.0
sqlalchemy>=2.0.25
alembic>=1.13.0
supabase>=2.3.0
greenlet>=3.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI/LLM
google-generativeai>=0.7.2
jinja2>=3.1.3

# Caching & Background Tasks
redis>=5.0.1
celery>=5.3.6

# Utilities
python-dotenv>=1.0.1
python-dateutil>=2.8.2
email-validator>=2.1.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=22.0.0
httpx>=0.26.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Monitoring & Logging
sentry-sdk[fastapi]>=1.40.0
